Figured out how to map temp data, now need to incorporate sparklines
