+++ package.json
@@ -16,6 +16,7 @@
     "next": "13.5.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-redux": "^9.1.2"
+    "react-redux": "^9.1.2",
+    "react-sparklines": "^1.7.0"
   }
 }
